rag-streamlit/requirements.txt
streamlit
openai>=1.0.0
faiss-cpu
numpy
PyPDF2
toml
